+++ src/main/java/com/rwtema/funkylocomotion/api/package-info.java
@@ -1,0 +1,4 @@
+@API(apiVersion = "2.0", owner = "funkylocomotion", provides = "funkylocomotion_api")
+package com.rwtema.funkylocomotion.api;
+
+import net.minecraftforge.fml.common.API;